Refactoring renderers
