+++ lib/core/src/control.rs
@@ -1,4 +1,4 @@
 //! Root module for all control abstractions, like event loops or event systems.
 
 pub mod callback;
-pub mod mouse_manager;
+pub mod io;
