+++ lib/core/src/control/io.rs
@@ -1,0 +1,3 @@
+//! Root module for all Input / Output events and devices.
+
+pub mod mouse;
