+++ lib/core/src/control/io/mouse.rs
@@ -1,6 +1,6 @@
 #![allow(missing_docs)]
 
-use crate::display::render::css3d::DOMContainer;
+use crate::system::web::dom::DOMContainer;
 use crate::system::web::dyn_into;
 use crate::system::web::Result;
 use crate::system::web::Error;
