+++ lib/core/src/display.rs
@@ -5,6 +5,5 @@ pub mod camera;
 pub mod shape;
 pub mod symbol;
 pub mod object;
-pub mod render;
 pub mod world;
 pub mod navigation;
