+++ lib/core/src/display/navigation/navigator.rs
@@ -6,9 +6,9 @@ use events::NavigatorEvents;
 use events::ZoomEvent;
 use events::PanEvent;
 use crate::system::web::Result;
-use crate::display::render::css3d::Camera;
-use crate::display::render::css3d::CameraType;
-use crate::display::render::css3d::DOMContainer;
+use crate::system::web::dom::Camera;
+use crate::system::web::dom::CameraType;
+use crate::system::web::dom::DOMContainer;
 use crate::animation::position::HasPosition;
 use crate::animation::physics::inertia::PhysicsSimulator;
 use crate::animation::physics::inertia::SpringProperties;
