+++ lib/core/src/display/navigation/navigator/events.rs
@@ -1,17 +1,19 @@
-use crate::control::mouse_manager::MouseManager;
-use crate::control::mouse_manager::MouseClickEvent;
-use crate::control::mouse_manager::MouseWheelEvent;
-use crate::control::mouse_manager::MousePositionEvent;
-use crate::control::mouse_manager::MouseButton;
-use crate::control::mouse_manager::WheelEventListener;
-use crate::control::mouse_manager::MouseEventListener;
+use crate::control::io::mouse::MouseManager;
+use crate::control::io::mouse::MouseClickEvent;
+use crate::control::io::mouse::MouseWheelEvent;
+use crate::control::io::mouse::MousePositionEvent;
+use crate::control::io::mouse::MouseButton;
+use crate::control::io::mouse::WheelEventListener;
+use crate::control::io::mouse::MouseEventListener;
 use crate::system::web::Result;
-use crate::display::render::css3d::DOMContainer;
+use crate::system::web::dom::DOMContainer;
 
 use nalgebra::Vector2;
 use std::rc::Rc;
 use std::cell::RefCell;
 use nalgebra::zero;
+
+
 
 // =================
 // === ZoomEvent ===
