+++ lib/core/src/system/web.rs
@@ -1,0 +1,5 @@
+//! Root module for web-based functionalities.
+
+pub mod dom;
+
+pub use basegl_system_web::*;
