+++ lib/core/src/system/web/dom/html/html_object.rs
@@ -2,7 +2,7 @@
 
 use crate::prelude::*;
 
-use crate::display::render::css3d::Object;
+use crate::system::web::dom::Object;
 use crate::system::web::create_element;
 use crate::system::web::dyn_into;
 use crate::system::web::Result;
