+++ lib/core/src/system/web/dom/object.rs
@@ -2,7 +2,7 @@
 
 use crate::prelude::*;
 
-use crate::display::render::css3d::Transform;
+use crate::system::web::dom::Transform;
 use crate::animation::position::HasPosition;
 
 use nalgebra::UnitQuaternion;
